move apis from hubd to userd
