+++ api/hubd/pb/hubd.proto
@@ -178,71 +178,6 @@ message DestroyAccountRequest {}
 
 message DestroyAccountResponse {}
 
-message ArchivesLsRequest {
-}
-
-message ArchivesLsResponse {
-    repeated ArchiveLsItem archives = 1;
-}
-
-message ArchiveLsItem {
-    string cid = 1;
-    repeated ArchiveLsItemMetadata info = 2;
-}
-
-message ArchiveLsItemMetadata {
-    uint64 deal_id = 1;
-}
-
-message ArchivesImportRequest {
-    string cid = 1;
-    repeated uint64 deal_ids = 2;
-}
-
-message ArchivesImportResponse {}
-
-message ArchiveRetrievalLsRequest {}
-
-message ArchiveRetrievalLsResponse {
-    repeated ArchiveRetrievalLsItem retrievals = 1;
-}
-
-message ArchiveRetrievalLsItem {
-	string id = 1;
-	string cid = 2;
-	ArchiveRetrievalStatus status = 3;
-	string failure_cause = 4;
-	int64 created_at = 5;
-
-	oneof retrieval_type {
-		ArchiveRetrievalLsItemNewBucket new_bucket = 6;
-	}
-}
-
-message ArchiveRetrievalLsItemNewBucket {
-	string name = 1;
-	bool private = 2;
-}
-
-enum ArchiveRetrievalStatus {
-    ARCHIVE_RETRIEVAL_STATUS_UNSPECIFIED = 0;
-    ARCHIVE_RETRIEVAL_STATUS_QUEUED = 1;
-    ARCHIVE_RETRIEVAL_STATUS_EXECUTING = 2;
-    ARCHIVE_RETRIEVAL_STATUS_MOVETOBUCKET = 3;
-    ARCHIVE_RETRIEVAL_STATUS_FAILED = 4;
-    ARCHIVE_RETRIEVAL_STATUS_CANCELED = 5;
-    ARCHIVE_RETRIEVAL_STATUS_SUCCESS = 6;
-}
-
-message ArchiveRetrievalLogsRequest {
-    string id = 1;
-}
-
-message ArchiveRetrievalLogsResponse {
-    string msg = 1;
-}
-
-
 service APIService {
     rpc BuildInfo(BuildInfoRequest) returns (BuildInfoResponse) {}
 
@@ -272,12 +207,4 @@ service APIService {
     rpc IsOrgNameAvailable(IsOrgNameAvailableRequest) returns (IsOrgNameAvailableResponse) {}
 
     rpc DestroyAccount(DestroyAccountRequest) returns (DestroyAccountResponse) {}
-
-    // Archives Import
-    rpc ArchivesLs(ArchivesLsRequest) returns (ArchivesLsResponse) {}
-    rpc ArchivesImport(ArchivesImportRequest) returns (ArchivesImportResponse) {}
-
-    // Archive Retrieval
-    rpc ArchiveRetrievalLs(ArchiveRetrievalLsRequest) returns (ArchiveRetrievalLsResponse){}
-    rpc ArchiveRetrievalLogs(ArchiveRetrievalLogsRequest) returns (stream ArchiveRetrievalLogsResponse) {}
 }
