+++ integrationtest/pg/pg_test.go
@@ -25,7 +25,8 @@ import (
 	pb "github.com/textileio/textile/v2/api/bucketsd/pb"
 	"github.com/textileio/textile/v2/api/common"
 	hc "github.com/textileio/textile/v2/api/hubd/client"
-	hubpb "github.com/textileio/textile/v2/api/hubd/pb"
+	uc "github.com/textileio/textile/v2/api/usersd/client"
+	userspb "github.com/textileio/textile/v2/api/usersd/pb"
 	"github.com/textileio/textile/v2/buckets/archive/retrieval"
 	"github.com/textileio/textile/v2/buckets/archive/tracker"
 	"github.com/textileio/textile/v2/core"
@@ -41,7 +42,7 @@ func TestMain(m *testing.M) {
 
 func TestCreateBucket(t *testing.T) {
 	powc, _ := StartPowergate(t, true)
-	ctx, _, _, client, _, shutdown := setup(t)
+	ctx, _, _, _, client, _, shutdown := setup(t)
 	defer shutdown()
 
 	// User is now created, so it should exist after spinup.
@@ -61,7 +62,7 @@ func TestCreateBucket(t *testing.T) {
 func TestArchiveTracker(t *testing.T) {
 	util.RunFlaky(t, func(t *util.FlakyT) {
 		_, _ = StartPowergate(t, true)
-		ctx, conf, _, client, repo, shutdown := setup(t)
+		ctx, conf, _, _, client, repo, shutdown := setup(t)
 
 		// Create bucket with a file.
 		b, err := client.Create(ctx)
@@ -107,7 +108,7 @@ func TestArchiveTracker(t *testing.T) {
 func TestArchiveBucketWorkflow(t *testing.T) {
 	util.RunFlaky(t, func(t *util.FlakyT) {
 		_, _ = StartPowergate(t, true)
-		ctx, _, hubClient, buckClient, _, shutdown := setup(t)
+		ctx, _, _, usersClient, buckClient, _, shutdown := setup(t)
 		defer shutdown()
 
 		// Create bucket with a file.
@@ -135,7 +136,7 @@ func TestArchiveBucketWorkflow(t *testing.T) {
 		require.NotEmpty(t, deal1.Miner)
 
 		// List archives.
-		as, err := hubClient.ArchivesLs(ctx)
+		as, err := usersClient.ArchivesLs(ctx)
 		require.NoError(t, err)
 		require.Len(t, as.Archives, 1)
 		require.Equal(t, rootCid1, as.Archives[0].Cid)
@@ -160,7 +161,7 @@ func TestArchiveBucketWorkflow(t *testing.T) {
 		require.NotEmpty(t, deal2.Miner)
 
 		// List archives.
-		as, err = hubClient.ArchivesLs(ctx)
+		as, err = usersClient.ArchivesLs(ctx)
 		require.NoError(t, err)
 		require.Len(t, as.Archives, 1)
 		require.Equal(t, rootCid2, as.Archives[0].Cid)
@@ -172,7 +173,7 @@ func TestArchiveBucketWorkflow(t *testing.T) {
 func TestArchivesLs(t *testing.T) {
 	util.RunFlaky(t, func(t *util.FlakyT) {
 		_, _ = StartPowergate(t, true)
-		ctx, _, hubClient, buckClient, _, shutdown := setup(t)
+		ctx, _, _, usersClient, buckClient, _, shutdown := setup(t)
 		defer shutdown()
 
 		// Create Bucket 1, and archive it.
@@ -189,7 +190,7 @@ func TestArchivesLs(t *testing.T) {
 		deal1 := res.Current.DealInfo[0]
 
 		// List archives, length should be 1.
-		as, err := hubClient.ArchivesLs(ctx)
+		as, err := usersClient.ArchivesLs(ctx)
 		require.NoError(t, err)
 		require.Len(t, as.Archives, 1)
 		require.Equal(t, rootCid1, as.Archives[0].Cid)
@@ -210,7 +211,7 @@ func TestArchivesLs(t *testing.T) {
 		deal2 := res.Current.DealInfo[0]
 
 		// List archives, length should be 2.
-		as, err = hubClient.ArchivesLs(ctx)
+		as, err = usersClient.ArchivesLs(ctx)
 		require.NoError(t, err)
 		require.Len(t, as.Archives, 2)
 		sort.Slice(as.Archives, func(i, j int) bool {
@@ -239,7 +240,7 @@ func TestArchivesLs(t *testing.T) {
 func TestArchivesImport(t *testing.T) {
 	util.RunFlaky(t, func(t *util.FlakyT) {
 		_, _ = StartPowergate(t, true)
-		hubClient, threadsclient, buckClient, conf, _, shutdown := createInfra(t)
+		hubClient, usersClient, threadsclient, buckClient, conf, _, shutdown := createInfra(t)
 		defer shutdown()
 
 		// Create an archive for account-1.
@@ -263,16 +264,16 @@ func TestArchivesImport(t *testing.T) {
 		ctxAccount2 := createAccount(t, hubClient, threadsclient, conf)
 
 		// Check no archives exist for account-2
-		as, err := hubClient.ArchivesLs(ctxAccount2)
+		as, err := usersClient.ArchivesLs(ctxAccount2)
 		require.NoError(t, err)
 		require.Len(t, as.Archives, 0)
 
 		// Import deal made by account-1.
-		err = hubClient.ArchivesImport(ctxAccount2, cid1, []uint64{deal.DealId})
+		err = usersClient.ArchivesImport(ctxAccount2, cid1, []uint64{deal.DealId})
 		require.NoError(t, err)
 
 		// Assert archives listing includes imported archive.
-		as, err = hubClient.ArchivesLs(ctxAccount2)
+		as, err = usersClient.ArchivesLs(ctxAccount2)
 		require.NoError(t, err)
 		require.Len(t, as.Archives, 1)
 
@@ -285,7 +286,7 @@ func TestArchivesImport(t *testing.T) {
 func TestArchiveUnfreeze(t *testing.T) {
 	util.RunFlaky(t, func(t *util.FlakyT) {
 		_, ipfsPow := StartPowergate(t, false)
-		hubClient, threadsclient, client, conf, _, shutdown := createInfra(t)
+		hubClient, usersClient, threadsclient, client, conf, _, shutdown := createInfra(t)
 		defer shutdown()
 		// Create an archive for account-1.
 		ctxAccount1 := createAccount(t, hubClient, threadsclient, conf)
@@ -314,11 +315,11 @@ func TestArchiveUnfreeze(t *testing.T) {
 		ctxAccount2 := createAccount(t, hubClient, threadsclient, conf)
 
 		// Obvious assertion about no existing retrievals.
-		rs, err := hubClient.ArchiveRetrievalLs(ctxAccount2)
+		rs, err := usersClient.ArchiveRetrievalLs(ctxAccount2)
 		require.NoError(t, err)
 		require.Len(t, rs.Retrievals, 0)
 		// Import deal made by account-1.
-		err = hubClient.ArchivesImport(ctxAccount2, ccid, []uint64{deal.DealId})
+		err = usersClient.ArchivesImport(ctxAccount2, ccid, []uint64{deal.DealId})
 		require.NoError(t, err)
 
 		// Delete `ccid` from go-ipfs of Powergate, and go-ipfs of Hub.
@@ -362,14 +363,14 @@ func TestArchiveUnfreeze(t *testing.T) {
 		buck, err = client.Create(ctxAccount2, c.WithName("super-bucket"), c.WithCid(ccid), c.WithUnfreeze(true))
 		require.NoError(t, err)
 		// Wait for the retrieval to finish.
-		var r *hubpb.ArchiveRetrievalLsItem
+		var r *userspb.ArchiveRetrievalLsItem
 		require.Eventually(t, func() bool {
-			rs, err = hubClient.ArchiveRetrievalLs(ctxAccount2)
+			rs, err = usersClient.ArchiveRetrievalLs(ctxAccount2)
 			require.NoError(t, err)
 			require.Len(t, rs.Retrievals, 1)
 			r = rs.Retrievals[0]
-			require.NotEqual(t, hubpb.ArchiveRetrievalStatus_ARCHIVE_RETRIEVAL_STATUS_FAILED, r.Status)
-			return r.Status == hubpb.ArchiveRetrievalStatus_ARCHIVE_RETRIEVAL_STATUS_SUCCESS
+			require.NotEqual(t, userspb.ArchiveRetrievalStatus_ARCHIVE_RETRIEVAL_STATUS_FAILED, r.Status)
+			return r.Status == userspb.ArchiveRetrievalStatus_ARCHIVE_RETRIEVAL_STATUS_SUCCESS
 		}, 2*time.Minute, 2*time.Second)
 		// Check if what we got from Filecoin is the same file that
 		// account-1 saved in the archived bucket.
@@ -392,7 +393,7 @@ func TestArchiveUnfreeze(t *testing.T) {
 		defer cancel()
 		ch := make(chan string, 100)
 		go func() {
-			err = hubClient.ArchiveRetrievalLogs(ctxAccount2, r.Id, ch)
+			err = usersClient.ArchiveRetrievalLogs(ctxAccount2, r.Id, ch)
 			close(ch)
 		}()
 		count := 0
@@ -411,7 +412,7 @@ func TestArchiveUnfreeze(t *testing.T) {
 func TestArchiveWatch(t *testing.T) {
 	util.RunFlaky(t, func(t *util.FlakyT) {
 		_, _ = StartPowergate(t, true)
-		ctx, _, _, client, _, shutdown := setup(t)
+		ctx, _, _, _, client, _, shutdown := setup(t)
 		defer shutdown()
 
 		b, err := client.Create(ctx)
@@ -445,7 +446,7 @@ func TestArchiveWatch(t *testing.T) {
 func TestFailingArchive(t *testing.T) {
 	util.RunFlaky(t, func(t *util.FlakyT) {
 		_, _ = StartPowergate(t, true)
-		ctx, _, _, client, _, shutdown := setup(t)
+		ctx, _, _, _, client, _, shutdown := setup(t)
 		defer shutdown()
 
 		b, err := client.Create(ctx)
@@ -507,14 +508,14 @@ func addDataFileToBucket(ctx context.Context, t util.TestingTWithCleanup, client
 	return strings.SplitN(root.String(), "/", 4)[2]
 }
 
-func setup(t util.TestingTWithCleanup) (context.Context, core.Config, *hc.Client, *c.Client, string, func()) {
-	hubClient, threadsclient, buckClient, conf, repo, shutdown := createInfra(t)
+func setup(t util.TestingTWithCleanup) (context.Context, core.Config, *hc.Client, *uc.Client, *c.Client, string, func()) {
+	hubClient, usersClient, threadsclient, buckClient, conf, repo, shutdown := createInfra(t)
 	ctx := createAccount(t, hubClient, threadsclient, conf)
 
-	return ctx, conf, hubClient, buckClient, repo, shutdown
+	return ctx, conf, hubClient, usersClient, buckClient, repo, shutdown
 }
 
-func createInfra(t util.TestingTWithCleanup) (*hc.Client, *tc.Client, *c.Client, core.Config, string, func()) {
+func createInfra(t util.TestingTWithCleanup) (*hc.Client, *uc.Client, *tc.Client, *c.Client, core.Config, string, func()) {
 	conf := apitest.DefaultTextileConfig(t)
 	conf.AddrPowergateAPI = powAddr
 	conf.ArchiveJobPollIntervalFast = time.Second * 5
@@ -532,10 +533,12 @@ func createInfra(t util.TestingTWithCleanup) (*hc.Client, *tc.Client, *c.Client,
 	})
 	hubclient, err := hc.NewClient(target, opts...)
 	require.NoError(t, err)
+	usersclient, err := uc.NewClient(target, opts...)
+	require.NoError(t, err)
 	threadsclient, err := tc.NewClient(target, opts...)
 	require.NoError(t, err)
 
-	return hubclient, threadsclient, client, conf, repo, shutdown
+	return hubclient, usersclient, threadsclient, client, conf, repo, shutdown
 
 }
